requirements.txt
boto3==1.26.102
celery==5.3.1
dj-database-url==0.5.0
django-ace==1.0.12
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.19.1
django-apptemplates==1.5
django-clearcache==1.2.1
django-colorfield==0.7.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==21.1
django-redis==5.2.0
django==4.0.5
djangorestframework==3.13.1
drf-api-logger==1.1.10
drf-yasg==1.21.3
geoip2==4.6.0
gunicorn==20.1.0
humanize==4.2.0
moviepy==1.0.3
paramiko==2.11.0
pdf2image==1.16.3
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
python-dateutil==2.8.2
redis==4.3.3
requests==2.28.0
tqdm==4.63.0
twilio==7.7.1
sentry-sdk==1.6.0
whitenoise==6.2.0